Fix type hint to be on par with spec
https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#cost-object

+++ openbrokerapi/catalog.py
@@ -39,7 +39,7 @@ class ServiceDashboardClient:
 
 class ServicePlanCost:
     def __init__(self,
-                 amount: Dict[str, int],
+                 amount: Dict[str, float],
                  unit: str,
                  **kwargs):
         self.amount = amount
